Project cleanup and fixes

+++ mooshak_testing/compare_files.sh
@@ -9,26 +9,35 @@ totalIncompatibilidades=0
 
 echo -e "\n"
 
-for ((meta=1; meta<=2; meta++)); do
+for ((meta=1; meta<=3; meta++)); do
     incompatibilidades=0
     
-    for ucfile in meta${meta}/*.uc; do
-        outfile="${ucfile%.uc}.out"
-        resfile="${ucfile%.uc}.res"
+    for ucfile in meta${meta}/*.{uc,c}; do
+        if [ -f "$ucfile" ]; then
+            if [[ "$ucfile" == *.uc ]]; then
+                outfile="${ucfile%.uc}.out"
+                resfile="${ucfile%.uc}.res"
+            elif [[ "$ucfile" == *.c ]]; then
+                outfile="${ucfile%.c}.out"
+                resfile="${ucfile%.c}.res"
+            fi
         
-        # Executar o uccompiler no arquivo .uc e gerar um arquivo .res
-        if [ "$meta" -eq 1 ]; then
-            ./uccompiler -l < "$ucfile" > "$resfile"
-        elif [ "$meta" -eq 2 ]; then
-            ./uccompiler -t < "$ucfile" > "$resfile"
-        fi
-        
-        # Usar diff para comparar .res e .out
-        if ! diff -q "$resfile" "$outfile" > /dev/null; then
-            echo "Erro > $ucfile"
-            # Incrementa o contador de incompatibilidades
-            ((incompatibilidades++))
-            ((totalIncompatibilidades++))
+            # Executar o uccompiler no arquivo .uc e gerar um arquivo .res
+            if [ "$meta" -eq 1 ]; then
+                ./uccompiler -l < "$ucfile" > "$resfile"
+            elif [ "$meta" -eq 2 ]; then
+                ./uccompiler -t < "$ucfile" > "$resfile"
+            elif [ "$meta" -eq 3 ]; then
+                ./uccompiler -s < "$ucfile" > "$resfile"
+            fi
+            
+            # Usar diff para comparar .res e .out
+            if ! diff -q "$resfile" "$outfile" > /dev/null; then
+                echo "Erro > $ucfile"
+                # Incrementa o contador de incompatibilidades
+                ((incompatibilidades++))
+                ((totalIncompatibilidades++))
+            fi
         fi
     done
     
@@ -38,6 +47,7 @@ for ((meta=1; meta<=2; meta++)); do
         echo -e "${YELLOW}Testes da meta${meta}: ${RED}NOK${NC}"
     fi
 done
+
 
 # Imprime o total de incompatibilidades encontradas
 if [ "$totalIncompatibilidades" -eq 0 ]; then
